requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
supabase==2.11.0
google-generativeai==0.8.4
langgraph==0.2.60
langchain-core==0.3.28
langchain-google-genai==2.0.8
httpx==0.28.1
numpy==2.2.1
